review: drop narration & redundant casts
- ChainSpecBasedSpecProvider: drop verbose BlockOf comment block
- HistoryPruner: drop misplaced field-level sentinel comment
- Merkleizer: (ulong)1 << k -> 1UL << k
- Test files: drop (ulong) casts where suffix or local const suffices

+++ src/Nethermind/Nethermind.TxPool.Test/TxPoolTests.Blobs.cs
@@ -326,7 +326,7 @@ namespace Nethermind.TxPool.Test
                 .WithShardBlobTxTypeAndFieldsIfBlobTx()
                 .WithMaxFeePerGas(1.GWei)
                 .WithMaxPriorityFeePerGas(1.GWei)
-                .WithNonce((ulong)2)
+                .WithNonce(2)
                 .SignedAndResolved(_ethereumEcdsa, TestItem.PrivateKeyA).TestObject;
 
             Assert.That(_txPool.SubmitTx(firstTx, TxHandlingOptions.None), Is.EqualTo(AcceptTxResult.Accepted));
